+++ src/main/java/org/springframework/social/partnercenter/api/billing/usage/UtilizationRecord.java
@@ -1,30 +1,31 @@
 package org.springframework.social.partnercenter.api.billing.usage;
 
 import java.math.BigDecimal;
-import java.util.Map;
+import java.time.ZonedDateTime;
 
-public class UtilizationRecord {
-	private String usageStartTime;
-	private String usageEndTime;
+import org.springframework.social.partnercenter.api.ResourceBase;
+
+public class UtilizationRecord extends ResourceBase {
+	private ZonedDateTime usageStartTime;
+	private ZonedDateTime usageEndTime;
 	private AzureResource resource;
 	private BigDecimal quantity;
 	private String unit;
 	private InfoFields infoFields;
-	private Map<String, String> attributes;
 
-	public String getUsageStartTime() {
+	public ZonedDateTime getUsageStartTime() {
 		return usageStartTime;
 	}
 
-	public void setUsageStartTime(String usageStartTime) {
+	public void setUsageStartTime(ZonedDateTime usageStartTime) {
 		this.usageStartTime = usageStartTime;
 	}
 
-	public String getUsageEndTime() {
+	public ZonedDateTime getUsageEndTime() {
 		return usageEndTime;
 	}
 
-	public void setUsageEndTime(String usageEndTime) {
+	public void setUsageEndTime(ZonedDateTime usageEndTime) {
 		this.usageEndTime = usageEndTime;
 	}
 
@@ -58,13 +59,5 @@ public class UtilizationRecord {
 
 	public void setInfoFields(InfoFields infoFields) {
 		this.infoFields = infoFields;
-	}
-
-	public Map<String, String> getAttributes() {
-		return attributes;
-	}
-
-	public void setAttributes(Map<String, String> attributes) {
-		this.attributes = attributes;
 	}
 }