+++ src/main/java/org/springframework/social/partnercenter/api/customer/user/AdminUserOperations.java
@@ -2,7 +2,6 @@ package org.springframework.social.partnercenter.api.customer.user;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.social.partnercenter.api.PartnerCenterResponse;
-import org.springframework.social.partnercenter.api.customer.Role;
 import org.springframework.social.partnercenter.api.customer.user.request.CustomerUserAssignLicenses;
 
 public interface AdminUserOperations extends UserOperations{
@@ -16,15 +15,11 @@ public interface AdminUserOperations extends UserOperations{
 	ResponseEntity<CustomerUser> updateUser(String customerTenantId, CustomerUser request, String userId);
 	ResponseEntity deleteUser(String customerTenantId, String userId);
 	ResponseEntity<CustomerUser> updateUserPassword(String customerTenantId, String userId, CustomerUser request);
-	ResponseEntity<PartnerCenterResponse<Role>> getUserRoles(String customerTenantId, String userId);
-	ResponseEntity<PartnerCenterResponse<Role>> getAllRoles(String customerTenantId);
-	ResponseEntity<PartnerCenterResponse<Role>> getRolesByRoleId(String customerTenantId, String RoleId);
-
 	/**
 	 * Gets a list of deleted CustomerUser resources for a customer by customer ID.
 	 *
 	 * @param customerId The value is a GUID formatted customer-id that identifies the customer.
-	 * @return List of deleted CustomerUsers
+	 * @return List of deleted {@link CustomerUser}
 	 */
 	ResponseEntity<PartnerCenterResponse<CustomerUser>> getDeletedUsers(String customerId);
 
@@ -33,7 +28,7 @@ public interface AdminUserOperations extends UserOperations{
 	 *
 	 * @param customerId The value is a GUID formatted customer-id that identifies the customer.
 	 * @param size The maximum size of the list to be returned
-	 * @return List of deleted CustomerUsers
+	 * @return List of deleted {@link CustomerUser}
 	 */
 	ResponseEntity<PartnerCenterResponse<CustomerUser>> getDeletedUsers(String customerId, Integer size);
 }