+++ src/test/java/org/springframework/social/partnercenter/test/stubs/StubURI.java
@@ -23,4 +23,8 @@ public class StubURI {
 				.build()
 				.toUri();
 	}
+
+	public static URI customersBase(int port) {
+		return baseURI(port, "v1", "customers");
+	}
 }